src/requirements.txt
python-telegram-bot==12.2.0
openpyxl==3.0.1
scrapinghub==2.3.0
falcon==2.0.0
PyYAML==5.3
pandas==0.25.3
envparse==0.2.0
sentry-sdk==0.13.5
msgpack==0.6.1
pytest==5.3.2
mixer==6.1.3
Faker==0.9.1